Remove URLs as possible Endpoint specifiers
Parsing URLs introduces a lot of complexity. Specifyng a URL also conflates
choice of endpoint with choice of protocol used to reach the endpoint (i.e.,
it confuses endpoints and transport parameters). The wording allows us to
add URLs back later if we want them, but I wonder if the right API for it
might be:

  preconnection := NewPreconnectionFromURL(url, localEndpoint);

rather than specifying the endpoint using a URL then passing that to
NewPreconnection() along with the transport and security parameters.