Proposed alternative intro text
I think the intro currently doesn't fit well the content of the document as any motivation for this work should be in the Arch doc. If we want to retain the text that I deleted we can maybe add it to the Intro or background section in the Arch doc...?

Note that the second paragraph I've added is stolen from the intro of the Arch doc.