Edits on Section on Handling Path Changes
This is a first attempt to work on the section "Handling Path Changes" (see issue #545), however, I think more work is needed.

To be honest I don't fully understand this part and it would be could if someone else could provide more input about the System Policy part:
"For protocols that do not support multipath or migration, the Protocol Instances should be informed of the path change, but should not be forcibly disconnected if the previously used path becomes unavailable. An exception to this case is if the System Policy changes to prohibit traffic from the Connection based on its properties, in which case the Protocol Stack should be disconnected."

Also the last part about pooled connection should say more but I'm not sure what. We currently have now only this sentence left:
"Pooled Connections {{pooled-connections}} may add or remove underlying transport connections in a similar manner."

And finally the originally issue #545 was also mentioning fallback which I didn't address yet but I have this feeling that that might need a separate section. Should I add a separate issue for that or is that already addressed somewhere and I missed it?