Called the probes "keep alive packets"
We needed a name for the transport packets sent to probe the path. I was reluctant to use "message", which STUN uses, because that sounded like something the App generated. I could have chosen "signal"  but that sounds like it is something else and so I chose "packet" (which SCTP and TCP uses) to make sure people realise this appears on the wire.