Update draft-ietf-taps-interface.md
Edit after TAPS Interim, removing the second configuration knob. 
This adds support. One knob enables this to be turned. The default is the system default for the transport. That default might be different for each transport - but that actually might be what is desired. 
A 'very large' value would disable this mechanism, and this text explicitly says greater than "conn-timeout".