Update draft-ietf-taps-impl.md
After reading #533,  I reviewed the latest text. I'm unsure I want to make changes to Arch/API: I see path as "network path", that is the set of hops through which the packet is forwarded (the local interface being the first local hop, and PvD information/PMTU/etc relating to something non-local).  This changes the examples using interface to align with this.