A Transport Services System
This document has been edited many ways, and this PR proposes a change in title and refinement in the use of the word 'architecture' to focus on the system specification, rather than the architectural principles. It does not intend to change the system in any way, this is intended as a language revision, to make it slightly less philosophical.