Make UTO more clearly a TCP example
We provide UTO here as an example of something that is offered by a specific transport. It is true that QUIC also includes a UTO - but we don't have a QUIC spec yet, and anyway other protocols may also define transport-specific options.  I suggest this is sufficiently niche that the example prevails - when you tear doen a QUIC connection might be different to a TCP connection, because of differences in transport. I therefore propose some minor teaks and think we ought not to say about QUIC here.